Replace integer types with int

+++ src/Library/Axe.cs
@@ -10,14 +10,14 @@ namespace Library
         /// <summary>
         /// The attack value of the axe
         /// </summary>
-        public ushort Attack { get; }
+        public int Attack { get; }
 
         /// <summary>
         /// The defense value of the axe
         /// </summary>
-        public ushort Defense { get; }
+        public int Defense { get; }
 
-        public Axe(ushort attack, ushort defense)
+        public Axe(int attack, int defense)
         {
             this.Attack = attack;
             this.Defense = defense;
